requirements-dev.txt
# Dev tools
black==23.12.0
flake8==7.0.0
pre-commit==3.3.3
pytest==9.0.2
pytest-mock==3.15.1
mypy==1.5.1
ipython==8.37.0
jupyterlab==4.4.5

# Base dependencies from Lambda
pandas>=2.0
numpy<2
pyarrow==11.0.0
pyyaml
xgboost==2.0.3
scikit-learn>=1.3
mlflow>=2.9
requests>=2.31
selenium==4.11.2
webdriver-manager==4.0.2
beautifulsoup4
boto3>=1.30
geopy
